chore(YandexMetrica): move sdk loader to separate file
