+++ src/integrations/YandexMetrica/browser.js
@@ -6,8 +6,8 @@ import {
   getHashFromArrayWithDuplicate,
 } from '../../utils/commonUtils';
 import { NAME } from './constants';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import { getDefinedTraits } from '../../utils/utils';
+import { loader } from './loader';
 
 class YandexMetrica {
   constructor(config, analytics, destinationInfo) {
@@ -30,35 +30,19 @@ class YandexMetrica {
   }
 
   loadScript() {
-    (function (m, e, t, r, i, k, a) {
-      m[i] =
-        m[i] ||
-        function () {
-          (m[i].a = m[i].a || []).push(arguments);
-        };
-      m[i].l = 1 * new Date();
-      for (var j = 0; j < document.scripts.length; j++) {
-        if (document.scripts[j].src === r) {
-          return;
-        }
-      }
-      (k = e.createElement(t)),
-        (a = e.getElementsByTagName(t)[0]),
-        (k.async = 1),
-        (k.src = r),
-        k.setAttribute('data-loader', LOAD_ORIGIN),
-        a.parentNode.insertBefore(k, a);
-    })(window, document, 'script', 'https://mc.yandex.ru/metrika/tag.js', 'ym');
-
-    ym(this.tagId, 'init', {
-      clickmap: this.clickMap,
-      trackLinks: this.trackLinks,
-      accurateTrackBounce: this.accurateTrackBounce,
-      webvisor: this.webvisor,
-      ecommerce: this.containerName,
-    });
-    window[`${this.containerName}`] = window[`${this.containerName}`] || [];
-    window[`${this.containerName}`].push({});
+    loader(
+      window,
+      document,
+      'script',
+      'https://mc.yandex.ru/metrika/tag.js',
+      'ym',
+      this.tagId,
+      this.clickMap,
+      this.trackLinks,
+      this.accurateTrackBounce,
+      this.webvisor,
+      this.containerName,
+    );
   }
 
   init() {
